requirements.txt
# Core ML / Model dependencies
torch>=2.2.0
transformers>=4.47.0
accelerate>=0.30.0

# Streamlit UI
streamlit>=1.35.0

# Image & Video processing
opencv-python>=4.9.0
Pillow>=10.3.0
numpy>=1.26.0

# DICOM support
pydicom>=2.4.4

# Hugging Face Hub (for gated model authentication)
huggingface_hub>=0.23.0
